+++ examples/resources/woodpecker_user/resource.tf
@@ -1,0 +1,3 @@
+resource "woodpecker_user" "user" {
+  login = "testuser"
+}